+++ model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	UserId   int32  `gorm:"unique; not null" json:"userId"`
+	UserId   int64  `gorm:"unique; not null" json:"userId"`
 	Username string `gorm:"unique; varchar(255); not null" json:"userName"`
 	Password string `gorm:"type:varchar(255); not null" json:"-"`
 	Email    string `gorm:"type:varchar(255); not null" json:"email"`
